# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "users",
        "description": "Documentation du CRUD avec Swagger",
        "version": "1.0"
    },
    "paths": {
...
        "/api/hosts": {
            "get": {
                "tags": [
                    "Hosts"
                ],
                "summary": "Lister tous les animateurs",
                "operationId": "61fb7cf6434cd6a1dc845f6473680f64",
                "responses": {
                    "200": {
                        "description": "Liste des animateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "role": {
                                                "type": "string",
                                                "example": "animateur"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Hosts"
                ],
                "summary": "Créer un nouvel animateur",
                "operationId": "834c934f594f7e461b9c57b4950e7b29",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Animateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "host": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/hosts/{id}": {
            "get": {
                "tags": [
                    "Hosts"
                ],
                "summary": "Afficher un animateur",
                "operationId": "6a15087196d0467f62bb946379e9543a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'animateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'animateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "animateur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cet utilisateur n’est pas un animateur"
                    }
                }
            },
            "put": {
                "tags": [
                    "Hosts"
                ],
                "summary": "Mettre à jour un animateur",
                "operationId": "7509fea05e56da500488b55d7f91cf03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'animateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Animateur mis à jour avec succès"
                    },
                    "404": {
                        "description": "Cet utilisateur n’est pas un animateur"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hosts"
                ],
                "summary": "Supprimer un animateur",
                "operationId": "b604ac9104ad9953d3abff9d9d9f98b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'animateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Animateur supprimé avec succès"
                    },
                    "404": {
                        "description": "Cet utilisateur n’est pas un animateur"
                    }
                }
            }
        },
...
    },
...
}